% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylCalculation.R
\name{MethylationClassSummary}
\alias{MethylationClassSummary}
\title{"MethylCalculation"}
\usage{
MethylationClassSummary(x)
}
\arguments{
\item{start_classes}{The START methylation classes}

\item{u}{The paramater that desctibing the methylation probablity on CpG site}

\item{d}{The paramater that desctibing the de-methylation probablity on 5mCpG site}

\item{p}{The paramater that desctibing the methylation probablity on semi-CpG site}

\item{total}{The total genes when do the calculation}

\item{loops}{The loop times for the calculation}

\item{cell_cycle}{The cell cycle times}
}
\value{
end_classes The END methylation classes

end_class_mean The average methylation level after cell division.
}
\description{
This .....
}
\examples{
MethylCalculation(start_classes,u,d,p)
}
