% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylCalculation.R
\name{MethylCalculation}
\alias{MethylCalculation}
\title{"MethylCalculation"}
\usage{
MethylCalculation(original_classes, u, d, p, cell_cycle = 1)
}
\arguments{
\item{original_classes}{The original methylation classes}

\item{u}{The paramater that describing the methylation probablity on CpG site}

\item{d}{The paramater that describing the de-methylation probablity on 5mCpG site}

\item{p}{The paramater that describing the methylation probablity on semi-CpG site}

\item{cell_cycle}{The cell cycle times}

\item{total}{The total genes when do the calculation}

\item{loops}{The loop times for the calculation}
}
\value{
terminational_classes The terminational methylation classes

average_methylation_level The average methylation level after [n] cell cycle(s).
}
\description{
Calculate the ratio of each methylation class after n cell cycle(s).
}
\details{
The transition matrix of this model describes the changes of DNA methylation during one cell cycle in three steps:
passive demethylation by DNA replication, active DNA methylation changes affected by DNA methylation-modifying enzymes
and DNA methylation combinations during homologous recombination.
}
\examples{
MethylCalculation(original_classes,u,d,p)
MethylCalculation(c(0.1,0.2,0.3,0.1,0.1),u=0.01,d=0.2,p=0.8,cell_cycle=1)
MethylCalculation(c(0.1,0.2,0.3,0.1,0.1),u=0.01,d=0.2,p=0.8,cell_cycle=10)
}
\references{
.
}
