% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterEstimation.R
\name{newton}
\alias{newton}
\title{"ParameterEstimation"}
\usage{
newton(func = objfun, x0, tol = 1e-05, n.max = 100, ...)
}
\arguments{
\item{observation_matrix}{The START methylation classes}

\item{iter}{the iteration times of the parameter estimation using newton's method}

\item{cell_cycle}{The cell cycle times}
}
\value{
end_classes The END methylation classes

end_class_mean The average methylation level after cell division.
}
\description{
This .....
}
\examples{
ParameterEstimation(observation_matrix,iter=50,cell_cycle=1)
}
