% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterEstimation.R
\name{ParameterEstimation}
\alias{ParameterEstimation}
\title{"ParameterEstimation"}
\usage{
ParameterEstimation(observation_matrix, iter = 50, cell_cycle = 1)
}
\arguments{
\item{observation_matrix}{The START methylation classes}

\item{iter}{the iteration times of the parameter estimation using newton's method}

\item{cell_cycle}{The cell cycle times}
}
\value{
end_classes The END methylation classes

end_class_mean The average methylation level after cell division.
}
\description{
This .....
}
\details{
This ...
}
\examples{
ParameterEstimation(observation_matrix,iter=50,cell_cycle=1)
observation_matrix <- matrix(c(0.9388,0.0952,0.0377,0,0,0.0497,0.5873,0.1887,0.0344,0.0149,0.0096,0.2381,0.4151,0.0653,0.0876,0.0019,0.0635,0.3396,0.6151,0.253,0,0.0159,0.0189,0.2852,0.6444),5,5,byrow=T)
ParameterEstimation(observation_matrix,iter=50,cell_cycle=1)
ParameterEstimation(observation_matrix,iter=100,cell_cycle=2)
ParameterEstimation(observation_matrix,iter=100,cell_cycle=10)
}
\references{
This ...
}
