% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylCalculation.R
\name{MethylCalculation}
\alias{MethylCalculation}
\title{"MethylCalculation"}
\usage{
MethylCalculation(start_classes, u, d, p, cell_cycle = 1)
}
\arguments{
\item{start_classes}{The START methylation classes}

\item{u}{The paramater that describing the methylation probablity on CpG site}

\item{d}{The paramater that describing the de-methylation probablity on 5mCpG site}

\item{p}{The paramater that describing the methylation probablity on semi-CpG site}

\item{cell_cycle}{The cell cycle times}

\item{total}{The total genes when do the calculation}

\item{loops}{The loop times for the calculation}
}
\value{
end_classes The END methylation classes

end_class_mean The average methylation level after cell division.
}
\description{
This .....
}
\details{
This ...
}
\examples{
MethylCalculation(start_classes,u,d,p)
MethylCalculation(c(0.1,0.2,0.3,0.1,0.1),u=0.01,d=0.2,p=0.8,cell_cycle=1)
MethylCalculation(c(0.1,0.2,0.3,0.1,0.1),u=0.01,d=0.2,p=0.8,cell_cycle=10)
}
\references{
This ...
}
